Rewrite proxy utility to work on arbitrary class
We will need to define a proxy for a function on WorldMorph in order to
support Ctrl/Cmd-Q keyboard shortcuts. This is also cleaner code anyhow
(it uses function scoping instead of the global namespace and is
arguably easier to understand).

This diff is pretty hard to read because of indentation changes. You can
get a cleaner diff by specifying -w in git-diff (or ?w=1 in the URL on
GitHub).